Metric Revamp and Mounting Bracket Redesign
Also made a folder for manufacturing jigs. Adjusted forward bulkhead for metric screws and for use with its manufacturing jig.
On the Airbrakes, I increased the thickness of the paddles, replaced the mounting bracket with its new version for 3D printing, and replaced hardware with metric equivalents (and sized relevant holes in parts accordingly).